move `LoggingLevel` enum to another package
