+++ samanthas_telegram_bot/api_clients/smalltalk/smalltalk_client.py
@@ -16,7 +16,7 @@ from samanthas_telegram_bot.api_clients.auxil.constants import (
     SMALLTALK_URL_GET_TEST,
     DataDict,
 )
-from samanthas_telegram_bot.api_clients.auxil.enums import LoggingLevel, SmalltalkTestStatus
+from samanthas_telegram_bot.api_clients.auxil.enums import SmalltalkTestStatus
 from samanthas_telegram_bot.api_clients.auxil.models import NotificationParams
 from samanthas_telegram_bot.api_clients.base.base_api_client import BaseApiClient
 from samanthas_telegram_bot.api_clients.base.exceptions import BaseApiClientError
@@ -30,6 +30,7 @@ from samanthas_telegram_bot.auxil.escape_for_markdown import escape_for_markdown
 from samanthas_telegram_bot.auxil.log_and_notify import logs
 from samanthas_telegram_bot.data_structures.constants import ALL_LEVELS
 from samanthas_telegram_bot.data_structures.context_types import CUSTOM_CONTEXT_TYPES
+from samanthas_telegram_bot.data_structures.enums import LoggingLevel
 from samanthas_telegram_bot.data_structures.models import SmalltalkResult
 
 load_dotenv()
