+++ samanthas_telegram_bot/conversation/auxil/callback_query_reply_sender.py
@@ -4,7 +4,6 @@ from typing import Union
 from telegram import CallbackQuery, InlineKeyboardButton, InlineKeyboardMarkup
 from telegram.constants import ParseMode
 
-from samanthas_telegram_bot.api_clients.auxil.enums import LoggingLevel
 from samanthas_telegram_bot.auxil.log_and_notify import logs
 from samanthas_telegram_bot.conversation.auxil.enums import (
     CommonCallbackData,
@@ -22,7 +21,7 @@ from samanthas_telegram_bot.data_structures.constants import (
     Locale,
 )
 from samanthas_telegram_bot.data_structures.context_types import CUSTOM_CONTEXT_TYPES
-from samanthas_telegram_bot.data_structures.enums import AgeRangeType, Role
+from samanthas_telegram_bot.data_structures.enums import AgeRangeType, LoggingLevel, Role
 from samanthas_telegram_bot.data_structures.models import AssessmentQuestion
