+++ samanthas_telegram_bot/conversation/callbacks/registration/common_main_flow.py
@@ -13,7 +13,6 @@ from telegram.constants import ParseMode
 from telegram.ext import ConversationHandler
 
 from samanthas_telegram_bot.api_clients import BackendClient
-from samanthas_telegram_bot.api_clients.auxil.enums import LoggingLevel
 from samanthas_telegram_bot.api_clients.smalltalk.smalltalk_client import SmallTalkClient
 from samanthas_telegram_bot.auxil.log_and_notify import logs
 from samanthas_telegram_bot.conversation.auxil.callback_query_reply_sender import (
@@ -31,7 +30,7 @@ from samanthas_telegram_bot.conversation.auxil.helpers import answer_callback_qu
 from samanthas_telegram_bot.conversation.auxil.message_sender import MessageSender
 from samanthas_telegram_bot.data_structures.constants import EMAIL_PATTERN, LOCALES, Locale
 from samanthas_telegram_bot.data_structures.context_types import CUSTOM_CONTEXT_TYPES
-from samanthas_telegram_bot.data_structures.enums import Role
+from samanthas_telegram_bot.data_structures.enums import LoggingLevel, Role
 
 
 async def start(update: Update, context: CUSTOM_CONTEXT_TYPES) -> int:
